requirements.txt
# CLI AI Agent dependencies
# Python 3.10+

# Core
openai>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
ddgs>=9.0.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
responses>=0.24.0
